refactor(shared): move functions repository to models
